weather_pipeline/etl/load.py
import os
from dotenv import load_dotenv
from sqlalchemy import create_engine
import pandas as pd

load_dotenv()

POSTGRES_USER = os.getenv("POSTGRES_USER")
POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
POSTGRES_DB = os.getenv("POSTGRES_DB")
POSTGRES_HOST = os.getenv("POSTGRES_HOST")
POSTGRES_PORT = os.getenv("POSTGRES_PORT")

DATABASE_URL = (
    f"postgresql://{POSTGRES_USER}:{POSTGRES_PASSWORD}"
    f"@{POSTGRES_HOST}:{POSTGRES_PORT}/{POSTGRES_DB}"
)

engine = create_engine(DATABASE_URL)

def load_weather(df: pd.DataFrame, table_name="weather_data"):
    try:
        df.to_sql(name=table_name, con=engine, if_exists="append", index=False, schema="etl")
        print(f"Loaded {len(df)} rows into '{table_name}'")
    except Exception as e:
        print(f"Error loading data: {e}")
